Update and add comments
